alternative volume computation, tree removal functionality
changed the computation of total tree volume for a tree class in such that the trees update() function already informs the model about changes in volume.

Added removal functionality in such a way that removing a tree will remove its entry in its neighbors tree-lookup-list. Introduced DistPair, a class that simply holds a reference to a tree and a double distance. Used DistPair instead of generic Pair, since we can take advantage of the equals function to remove a distPair from a list.